Add MIT License to the project
This commit introduces the MIT License to the SQL Dump Splitter project. By including the MIT License, we clearly define the terms under which the project can be freely used, modified, and shared. The license fosters open collaboration and contribution while providing legal clarity and protection for both the contributors and users of the software.